requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio==4.2.0
appnope==0.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
babel==2.11.0
backcall==0.1.0
beautifulsoup4==4.12.2
bleach==4.1.0
# brotli-python
bzip2==1.0.8
ca-certificates==2024.3.11
cachetools==4.1.0
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.4
comm==0.2.1
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
distro==1.5.0
exceptiongroup==1.2.0
executing==0.8.3
filelock==3.0.12
gdown==3.11.0
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.14.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
httplib2==0.17.3
idna==2.9
ipykernel==6.28.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
jinja2==3.1.4
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_events==0.8.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab_pygments==0.1.2
jupyterlab_server==2.25.1
libcxx==14.0.6
libffi==3.4.4
libsodium==1.0.18
markupsafe==2.1.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
ncurses==6.4
nest-asyncio==1.6.0
notebook-shim==0.2.3
numpy==1.26.4
oauthlib==3.1.0
openssl==3.0.13
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==7.1.2
pip==24.0
platformdirs==3.10.0
prometheus_client==0.14.1
prompt-toolkit==3.0.5
prompt_toolkit==3.0.43
protobuf==3.11.3
psutil==5.9.0
ptyprocess==0.6.0
pure_eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.9.7
pygments==2.6.1
pysocks==1.7.1
python==3.10.14
python-dateutil==2.9.0post0
python-fastjsonschema==2.16.2
python-json-logger==2.0.7
pytz==2020.1
pyyaml==6.0.1
pyzmq==25.1.2
readline==8.2
referencing==0.30.2
requests==2.23.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
rsa==4.0
send2trash==1.8.2
setuptools==69.5.1
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.1.0
soupsieve==2.5
sqlite==3.45.3
stack_data==0.2.0
terminado==0.17.1
tinycss2==1.2.1
tk==8.6.14
tomli==2.0.1
tornado==6.3.3
tqdm==4.46.0
traitlets==4.3.3
typing-extensions==4.11.0
typing_extensions==4.11.0
tzdata==2024.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
wand==0.5.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.43.0
xlrd==1.2.0
xz==5.4.6
yaml==0.2.5
zeromq==4.3.5
zlib==1.2.13
